moved configuration files inside each container #6
